+++ Toggl.Api/DataObjects/Task.cs
@@ -9,19 +9,19 @@ namespace Toggl.Api.DataObjects
 	public class Task : BaseDataObject
 	{
 		[JsonProperty(PropertyName = "id")]
-		public int? Id { get; set; }
+		public long? Id { get; set; }
 
 		[JsonProperty(PropertyName = "name")]
 		public string? Name { get; set; }
 
 		[JsonProperty(PropertyName = "pid")]
-		public int? ProjectId { get; set; }
+		public long? ProjectId { get; set; }
 
 		[JsonProperty(PropertyName = "wid")]
-		public int? WorkspaceId { get; set; }
+		public long? WorkspaceId { get; set; }
 
 		[JsonProperty(PropertyName = "uid")]
-		public int? UserId { get; set; }
+		public long? UserId { get; set; }
 
 		[JsonProperty(PropertyName = "estimated_seconds")]
 		public int? EstimatedSeconds { get; set; }
