+++ Toggl.Api/Interfaces/IWorkspaceServiceAsync.cs
@@ -18,43 +18,43 @@ namespace Toggl.Api.Interfaces
 		/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/workspaces.md#get-workspace-users
 		/// </summary>
 		/// <param name="workspaceId"></param>
-		Task<List<User>> GetUsersAsync(int workspaceId);
+		Task<List<User>> GetUsersAsync(long workspaceId);
 
 		/// <summary>
 		/// Project users
 		/// </summary>
 		/// <param name="workspaceId"></param>
-		Task<List<ProjectUser>> GetProjectUsersAsync(int workspaceId);
+		Task<List<ProjectUser>> GetProjectUsersAsync(long workspaceId);
 
 		/// <summary>
 		/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/workspaces.md#get-workspace-clients
 		/// </summary>
 		/// <param name="workspaceId"></param>
-		Task<List<Client>> GetClientsAsync(int workspaceId);
+		Task<List<Client>> GetClientsAsync(long workspaceId);
 
 		/// <summary>
 		/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/workspaces.md#get-workspace-projects
 		/// </summary>
 		/// <param name="workspaceId"></param>
-		Task<List<Project>> GetProjectsAsync(int workspaceId);
+		Task<List<Project>> GetProjectsAsync(long workspaceId);
 
 		/// <summary>
 		/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/workspaces.md#get-workspace-tasks
 		/// </summary>
 		/// <param name="workspaceId"></param>
-		Task<List<DataObjects.Task>> GetTasksAsync(int workspaceId);
+		Task<List<DataObjects.Task>> GetTasksAsync(long workspaceId);
 
 		/// <summary>
 		/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/workspaces.md#get-workspace-tags
 		/// </summary>
 		/// <param name="workspaceId"></param>
-		Task<List<Tag>> GetTagsAsync(int workspaceId);
+		Task<List<Tag>> GetTagsAsync(long workspaceId);
 
 		/// <summary>
 		/// https://github.com/toggl/toggl_api_docs/blob/master/chapters/workspaces.md#get-workspace-timeentries
 		/// </summary>
 		/// <param name="workspaceId"></param>
 		/// <param name="workspaceTimeEntryId"></param>
-		Task<TimeEntry> GetTimeEntryAsync(int workspaceId, long workspaceTimeEntryId);
+		Task<TimeEntry> GetTimeEntryAsync(long workspaceId, long workspaceTimeEntryId);
 	}
 }
