+++ Toggl.Api/Services/WorkspaceServiceAsync.cs
@@ -30,49 +30,49 @@ namespace Toggl.Api.Services
 			return response.GetData<List<Workspace>>();
 		}
 
-		public async Task<List<User>> GetUsersAsync(int workspaceId)
+		public async Task<List<User>> GetUsersAsync(long workspaceId)
 		{
 			var url = string.Format(ApiRoutes.Workspace.ListWorkspaceUsersUrl, workspaceId);
 			var response = await TogglSrv.GetAsync(url).ConfigureAwait(false);
 			return response.GetData<List<User>>();
 		}
 
-		public async Task<List<ProjectUser>> GetProjectUsersAsync(int workspaceId)
+		public async Task<List<ProjectUser>> GetProjectUsersAsync(long workspaceId)
 		{
 			var url = string.Format(ApiRoutes.Workspace.ListWorkspaceProjectUsersUrl, workspaceId);
 			var response = await TogglSrv.GetAsync(url).ConfigureAwait(false);
 			return response.GetData<List<ProjectUser>>();
 		}
 
-		public async Task<List<Client>> GetClientsAsync(int workspaceId)
+		public async Task<List<Client>> GetClientsAsync(long workspaceId)
 		{
 			var url = string.Format(ApiRoutes.Workspace.ListWorkspaceClientsUrl, workspaceId);
 			var response = await TogglSrv.GetAsync(url).ConfigureAwait(false);
 			return response.GetData<List<Client>>();
 		}
 
-		public async Task<List<Project>> GetProjectsAsync(int workspaceId)
+		public async Task<List<Project>> GetProjectsAsync(long workspaceId)
 		{
 			var url = string.Format(ApiRoutes.Workspace.ListWorkspaceProjectsUrl, workspaceId);
 			var response = await TogglSrv.GetAsync(url).ConfigureAwait(false);
 			return response.GetData<List<Project>>();
 		}
 
-		public async Task<List<DataObjects.Task>> GetTasksAsync(int workspaceId)
+		public async Task<List<DataObjects.Task>> GetTasksAsync(long workspaceId)
 		{
 			var url = string.Format(ApiRoutes.Workspace.ListWorkspaceTasksUrl, workspaceId);
 			var response = await TogglSrv.GetAsync(url).ConfigureAwait(false);
 			return response.GetData<List<DataObjects.Task>>();
 		}
 
-		public async Task<List<Tag>> GetTagsAsync(int workspaceId)
+		public async Task<List<Tag>> GetTagsAsync(long workspaceId)
 		{
 			var url = string.Format(ApiRoutes.Workspace.ListWorkspaceTagsUrl, workspaceId);
 			var response = await TogglSrv.GetAsync(url).ConfigureAwait(false);
 			return response.GetData<List<Tag>>();
 		}
 
-		public async Task<TimeEntry> GetTimeEntryAsync(int workspaceId, long workspaceTimeEntryId)
+		public async Task<TimeEntry> GetTimeEntryAsync(long workspaceId, long workspaceTimeEntryId)
 		{
 			var url = string.Format(ApiRoutes.Workspace.GetWorkspaceTimeEntry, workspaceId, workspaceTimeEntryId);
 			var response = await TogglSrv.GetAsync(url).ConfigureAwait(false);
@@ -80,7 +80,7 @@ namespace Toggl.Api.Services
 			return data;
 		}
 
-		Task<List<DataObjects.Task>> IWorkspaceServiceAsync.GetTasksAsync(int workspaceId)
+		Task<List<DataObjects.Task>> IWorkspaceServiceAsync.GetTasksAsync(long workspaceId)
 			=> throw new NotSupportedException();
 	}
 }
